+++ main.py
@@ -10,7 +10,6 @@ from threading import Lock
 from pydantic import ValidationError
 
 from kytos.core import KytosNApp, log, rest
-from kytos.core.common import EntityStatus
 from kytos.core.helpers import (alisten_to, listen_to, load_spec,
                                 validate_openapi)
 from kytos.core.interface import TAG, UNI
@@ -18,12 +17,12 @@ from kytos.core.link import Link
 from kytos.core.rest_api import (HTTPException, JSONResponse, Request,
                                  get_json_or_400)
 from napps.kytos.mef_eline import controllers, settings
-from napps.kytos.mef_eline.exceptions import InvalidPath
+from napps.kytos.mef_eline.exceptions import DisabledSwitch, InvalidPath
 from napps.kytos.mef_eline.models import (EVC, DynamicPathManager, EVCDeploy,
                                           Path)
 from napps.kytos.mef_eline.scheduler import CircuitSchedule, Scheduler
-from napps.kytos.mef_eline.utils import (aemit_event, emit_event,
-                                         map_evc_event_content)
+from napps.kytos.mef_eline.utils import (aemit_event, check_disabled_component,
+                                         emit_event, map_evc_event_content)
 
 
 # pylint: disable=too-many-public-methods
@@ -233,11 +232,11 @@ class Main(KytosNApp):
             log.debug("create_circuit result %s %s", exception, 400)
             raise HTTPException(400, detail=str(exception)) from exception
 
-        if evc.is_intra_switch():
-            try:
-                self.check_disabled_component(evc.uni_a, evc.uni_z)
-            except HTTPException as exception:
-                raise HTTPException(
+        try:
+            check_disabled_component(evc.uni_a, evc.uni_z)
+        except DisabledSwitch as exception:
+            log.debug("create_circuit result %s %s", exception, 409)
+            raise HTTPException(
                     409,
                     detail=f"Path is not valid: {exception}"
                 ) from exception
@@ -349,6 +348,12 @@ class Main(KytosNApp):
             log.error(exception)
             log.debug("update result %s %s", exception, 400)
             raise HTTPException(400, detail=str(exception)) from exception
+        except DisabledSwitch as exception:
+            log.debug("update result %s %s", exception, 409)
+            raise HTTPException(
+                    409,
+                    detail=f"Path is not valid: {exception}"
+                ) from exception
 
         if evc.is_active():
             if enable is False:  # disable if active
@@ -991,28 +996,6 @@ class Main(KytosNApp):
                 evc = self._evc_from_dict(circuit)
                 self.circuits[c_id] = evc
         return self.circuits
-
-    @staticmethod
-    def check_disabled_component(uni_a: object, uni_z: object):
-        """Check if a switch or an interface is disabled"""
-        if uni_a.interface.switch.status == EntityStatus.DISABLED:
-            dpid = uni_a.interface.switch.dpid
-            raise HTTPException(
-                409,
-                detail=f"Switch {dpid} is disabled"
-            )
-        if uni_a.interface.status == EntityStatus.DISABLED:
-            id_ = uni_a.interface.id
-            raise HTTPException(
-                409,
-                detail=f"Interface {id_} is disabled"
-            )
-        if uni_z.interface.status == EntityStatus.DISABLED:
-            id_ = uni_z.interface.id
-            raise HTTPException(
-                409,
-                detail=f"Interface {id_} is disabled"
-            )
 
     # pylint: disable=attribute-defined-outside-init
     @alisten_to("kytos/of_multi_table.enable_table")